dismantle fwd.hpp for the most part
Having forward declarations of many types causes more problems than it solves.
It runs danger of IFNDR due to running SFINAE on incomplete types, and adds maintenance burden.
Many of the forward declarations are unnecessary, especially for small enumerations in their own header.

This PR also fixes some potential bugs in ast.hpp caused by forming std::span<T> where T is completed later.
