+++ include/cowel/ast_fwd.hpp
@@ -1,0 +1,14 @@
+#ifndef COWEL_AST_FWD_HPP
+#define COWEL_AST_FWD_HPP
+
+namespace cowel::ast {
+
+struct Markup_Element;
+struct Member_Value;
+struct Directive;
+struct Group_Member;
+struct Primary;
+
+} // namespace cowel::ast
+
+#endif
